mx62 - select appropriate subbatch unit when you change consumable type
The ingredients_prefill in ConsumableType is now prefill_data, and it includes
the sub batch unit (or nil).
When you add a new subbatch, it will have the unit from the last subbatch in the on-screen table
if such a thing exists, otherwise it will get the default unit from the prefill data.
When you switch consumable type, the subbatches on screen will be switched to the prefill unit,
ONLY if the prefill unit is non-null and none of the subbatches have numbers filled in.
This also covers the case where you are editing an existing batch: the unit shouldn't switch
of things that already exist.