mx65 - raise a support request from mixtio
On the view batch screen there is now a button to go to the support screen
(for that batch).
You can fill in the fields describing the problem.
Then press a button to open your email client with the details filled in.
Currently the email is directed to psd-help, because there is no
Mixtio queue.